reference_only/kinetic-scheme-visualizer/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kineticschemevisualizer"
version = "0.0.1"
dependencies = [
  "anywidget==0.9.13",
  "networkx==3.3",
  "pydantic==2.7.1",
  "pyglotaran==0.7.2",
  "xarray==2024.6.0",
]
readme = "README.md"

[project.optional-dependencies]
dev = ["watchfiles", "jupyterlab"]

# automatically add the dev feature to the default env (e.g., hatch shell)
[tool.hatch.envs.default]
python = "3.10"
features = ["dev"]

[tool.hatch.build]
only-packages = true
artifacts = ["kineticschemevisualizer/static/*"]

[tool.hatch.build.hooks.jupyter-builder]
build-function = "hatch_jupyter_builder.npm_builder"
ensured-targets = ["kineticschemevisualizer/static/widget.js"]
skip-if-exists = ["kineticschemevisualizer/static/widget.js"]
dependencies = ["hatch-jupyter-builder>=0.5.0"]

[tool.hatch.build.hooks.jupyter-builder.build-kwargs]
npm = "npm"
build_cmd = "build"
